feat(ocpp): add force push option to ocpp_push_request
This commit introduces a new parameter to the `ocpp_push_request` function
that allows forcing a request to be pushed even if the queue is full. If
the queue is full and the force option is set, the oldest request will be
dropped to make room for the new one. However, if the oldest request is
StartTransaction, StopTransaction or BootNotification, the next oldest
request will be dropped, which means those messages never get dropped.
This feature is useful in situations where it's critical to send a new
request regardless of the queue state.

BREAKING CHANGE: The `ocpp_push_request` function now requires an additional
boolean parameter. Existing calls to this function will need to be updated.